requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis and task queue
redis==5.0.1
celery==5.3.4

# Microsoft Bot Framework
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0

# HTTP clients
httpx==0.25.2
# aiohttp version will be installed automatically by botbuilder-integration-aiohttp

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data validation
pydantic==2.5.0
pydantic-settings==2.0.3

# AI and embeddings
anthropic==0.7.8
openai==1.3.7
sentence-transformers==2.2.2
qdrant-client==1.6.9
numpy==1.24.4

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1